refactor: signal handling
